requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
boto3
django-cors-headers
psycopg2-binary
gunicorn
python-dotenv
Pillow
moviepy
psutil
